Use size_t instead of unsigned

+++ stingraykit/metaprogramming/TypeTraits.h
@@ -59,7 +59,7 @@ namespace stingray
 
 // Array
 	template < typename T > struct IsArray											{ static const bool Value = false; };
-	template < typename T, unsigned N > struct IsArray<T[N]>						{ static const bool Value = true; };
+	template < typename T, size_t N > struct IsArray<T[N]>							{ static const bool Value = true; };
 
 	template < typename T > struct RemoveExtent										{ typedef T ValueT; };
 	template < typename T > struct RemoveExtent<T[]>								{ typedef T ValueT; };
